name yells if no name (solve the clear)
4.2.8

if(!taskNameInput) {
        alert("You need to fill in your initials!");
        return false;
    }
this yells at you if you did not type in your name at the end of the game.

 nameButtonEl.reset();
this is intended to reset the text box.  But I am only now just realizing I don't need this at all.  Once you type your name you are good.  I'll clear it tomorrow.